+++ src/main/java/luna/exception/LunaException.java
@@ -1,3 +1,4 @@
+package luna.exception;
 /**
  * Custom exception for Luna
  */
