+++ src/main/java/luna/task/DeadlineTask.java
@@ -1,7 +1,10 @@
+package luna.task;
 import java.time.LocalDate;
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
 import java.time.format.DateTimeParseException;
+
+import luna.exception.LunaException;
 
 /**
  * Represents a Deadline task with a description, end date and task type.
