+++ src/main/java/luna/task/Task.java
@@ -1,3 +1,4 @@
+package luna.task;
 /**
  * Task with a description and completion status.
  */
